Add `SufficientStatistics`
