+++ src/estimator/bayesian.rs
@@ -1,15 +1,10 @@
 use ndarray::prelude::*;
 
-use super::CPDEstimator;
-use crate::{
-    data::{CategoricalData, Data},
-    distribution::CategoricalCPD,
-    types::FxIndexSet,
-    utils::RMI,
-};
+use super::{CPDEstimator, CSSEstimator, SSE};
+use crate::{data::CategoricalData, distribution::CategoricalCPD};
 
 /// A struct representing a Bayesian estimator.
-#[derive(Clone, Debug)]
+#[derive(Clone, Copy, Debug, Default)]
 pub struct BayesianEstimator<Pi> {
     prior: Pi,
 }
@@ -48,52 +43,18 @@ impl<Pi> BayesianEstimator<Pi> {
 // NOTE: The prior is expressed as a scalar, which is the alpha for the Dirichlet distribution.
 impl CPDEstimator<CategoricalData, CategoricalCPD> for BE<f64> {
     fn fit(&self, data: &CategoricalData, x: usize, z: &[usize]) -> CategoricalCPD {
-        // Concat the variables to fit.
-        let x_z: FxIndexSet<_> = [x].iter().chain(z).cloned().collect();
+        // Get states and cardinality.
+        let (states, cards) = (data.states(), data.cardinality());
 
-        // Assert X_Z does not contain duplicates.
-        assert_eq!(x_z.len(), 1 + z.len(), "Variables to fit must be unique.");
-
-        // Get the reference to the labels, states and cardinality.
-        let (labels, states, cards) = (data.labels(), data.states(), data.cardinality());
-
-        // Assert the variables to fit are in the data.
-        assert!(
-            x_z.iter().all(|&i| i < labels.len()),
-            "Variables to fit must be in the data."
-        );
-
-        // Initialize ravel multi index.
-        let rmi = RMI::new(z.iter().map(|&i| cards[i]));
-        // Initialize the joint counts.
-        let mut n_xz: Array2<usize> = Array::zeros((rmi.cardinality().product(), cards[x]));
-
-        // Count the occurrences of the states.
-        data.values().rows().into_iter().for_each(|row| {
-            // Get the value of X as index.
-            let idx_x = row[x] as usize;
-            // Get the value of Z as index using the strides.
-            let idx_z = rmi.ravel(z.iter().map(|&i| row[i] as usize));
-            // Increment the joint counts.
-            n_xz[[idx_z, idx_x]] += 1;
-        });
-
-        // Marginalize the counts.
-        let n_z = n_xz.sum_axis(Axis(1));
-        // Compute the sample size.
-        let n = n_z.sum();
-
-        // Cast the counts to floating point.
-        let n_xz = n_xz.mapv(|x| x as f64);
-        let n_z = n_z.mapv(|x| x as f64);
-
-        // Align the marginal counts axes.
-        let n_z = n_z.insert_axis(Axis(1));
+        // Initialize the sufficient statistics estimator.
+        let sse = SSE::new();
+        // Compute sufficient statistics.
+        let (n_xz, n_z, n) = sse.fit(data, x, z);
 
         // Get the prior, as the alpha of the Dirichlet distribution.
         let alpha = *self.prior();
         // Compute the parameters by normalizing the counts with the prior.
-        let parameters = (&n_xz + alpha) / (n_z + alpha * cards[x] as f64);
+        let parameters = (&n_xz + alpha) / (n_z.insert_axis(Axis(1)) + alpha * cards[x] as f64);
 
         // Set the sample size.
         let sample_size = Some(n);
